Fix language switcher by centralizing state
The previous implementation of the `useTranslation` hook maintained its
own local state for the current language and translations. This caused
a bug where changing the language in one component (the language
switcher) did not update any other components on the page.

This commit resolves the issue by centralizing the translation state
using a React Context.

- A `TranslationProvider` is created to manage the global language
  state (`currentLanguage`, `translations`, `isLoading`).
- The `useTranslation` hook is refactored to consume this shared
  context, ensuring all components access the same state.
- The main application is wrapped in the `TranslationProvider`.
- A bug in the `HeroSection` where some text was using a legacy
  `data-i18n` attribute instead of the `t()` function was also fixed.

The `toggleLanguage` function in the new provider continues to dispatch
the `languageChanged` event to maintain compatibility with existing
non-React scripts.
